+++ include/domus/core/graph/graph_utilities.hpp
@@ -7,13 +7,13 @@
 
 class NodesContainer : protected IntHashSet {
   public:
-    void add_node(int node_id);
-    bool has_node(int node_id) const;
-    int get_one_node_id() const;
+    void add_node(size_t node_id);
+    bool has_node(size_t node_id) const;
+    size_t get_one_node_id() const;
     size_t size() const;
     bool empty() const;
-    void erase(int node_id);
-    void for_each(std::function<void(int)> func) const;
+    void erase(size_t node_id);
+    void for_each(std::function<void(size_t)> func) const;
 };
 
 class I_AdjacencyList;
@@ -27,9 +27,9 @@ class AdjacencyList {
     AdjacencyList(AdjacencyList&&) noexcept;
     AdjacencyList& operator=(AdjacencyList&&) noexcept;
 
-    void add_edge(int from_id, int to_id);
-    bool has_edge(int from_id, int to_id) const;
-    const NodesContainer& get_neighbors_of_node(int node_id) const;
-    void erase_edge(int from_id, int to_id);
-    void erase_node(int node_id);
+    void add_edge(size_t from_id, size_t to_id);
+    bool has_edge(size_t from_id, size_t to_id) const;
+    const NodesContainer& get_neighbors_of_node(size_t node_id) const;
+    void erase_edge(size_t from_id, size_t to_id);
+    void erase_node(size_t node_id);
 };
