+++ include/domus/orthogonal/drawing_stats.hpp
@@ -4,30 +4,30 @@
 
 struct OrthogonalDrawing;
 
-int compute_total_edge_length(const OrthogonalDrawing& result);
+size_t compute_total_edge_length(const OrthogonalDrawing& result);
 
-int compute_max_edge_length(const OrthogonalDrawing& result);
+size_t compute_max_edge_length(const OrthogonalDrawing& result);
 
 double compute_edge_length_std_dev(const OrthogonalDrawing& result);
 
-int compute_total_bends(const OrthogonalDrawing& result);
+size_t compute_total_bends(const OrthogonalDrawing& result);
 
-int compute_max_bends_per_edge(const OrthogonalDrawing& result);
+size_t compute_max_bends_per_edge(const OrthogonalDrawing& result);
 
 double compute_bends_std_dev(const OrthogonalDrawing& result);
 
-int compute_total_area(const OrthogonalDrawing& result);
+size_t compute_total_area(const OrthogonalDrawing& result);
 
-int compute_total_crossings(const OrthogonalDrawing& result);
+size_t compute_total_crossings(const OrthogonalDrawing& result);
 
 struct OrthogonalStats {
-    int crossings;
-    int bends;
-    int area;
-    int total_edge_length;
-    int max_edge_length;
+    size_t crossings;
+    size_t bends;
+    size_t area;
+    size_t total_edge_length;
+    size_t max_edge_length;
     double edge_length_stddev;
-    int max_bends_per_edge;
+    size_t max_bends_per_edge;
     double bends_stddev;
 };
