+++ src/core/containers.cpp
@@ -6,10 +6,10 @@
 #include <unordered_set>
 
 struct int_pair_hash {
-    int operator()(const std::pair<int, int>& p) const {
-        const int h1 = static_cast<int>(std::hash<int>{}(p.first));
-        const int h2 = static_cast<int>(std::hash<int>{}(p.second));
-        const int mult = h2 * static_cast<int>(0x9e3779b9);
+    size_t operator()(const std::pair<size_t, size_t>& p) const {
+        const size_t h1 = static_cast<size_t>(std::hash<size_t>{}(p.first));
+        const size_t h2 = static_cast<size_t>(std::hash<size_t>{}(p.second));
+        const size_t mult = h2 * static_cast<size_t>(0x9e3779b9);
         return h1 ^ (mult + (h1 << 6) + (h1 >> 2));
     }
 };
@@ -18,7 +18,7 @@ struct int_pair_hash {
 
 class I_IntPairHashSet {
   public:
-    std::unordered_set<std::pair<int, int>, int_pair_hash> pairs;
+    std::unordered_set<std::pair<size_t, size_t>, int_pair_hash> pairs;
 };
 
 PairIntHashSet::PairIntHashSet() { m_impl = std::make_unique<I_IntPairHashSet>(); }
@@ -29,12 +29,12 @@ PairIntHashSet::PairIntHashSet(PairIntHashSet&&) noexcept = default;
 
 PairIntHashSet& PairIntHashSet::operator=(PairIntHashSet&&) noexcept = default;
 
-void PairIntHashSet::add(int value_1, int value_2) {
+void PairIntHashSet::add(size_t value_1, size_t value_2) {
     assert(!has(value_1, value_2) && "PairIntHashSet::add: pair already exists");
     m_impl->pairs.insert({value_1, value_2});
 }
 
-bool PairIntHashSet::has(int value_1, int value_2) const {
+bool PairIntHashSet::has(size_t value_1, size_t value_2) const {
     return m_impl->pairs.contains({value_1, value_2});
 }
 
@@ -42,12 +42,12 @@ size_t PairIntHashSet::size() const { return m_impl->pairs.size(); }
 
 bool PairIntHashSet::empty() const { return m_impl->pairs.empty(); }
 
-void PairIntHashSet::erase(int value_1, int value_2) {
+void PairIntHashSet::erase(size_t value_1, size_t value_2) {
     assert(has(value_1, value_2) && "PairIntHashSet::erase: pair does not exist");
     m_impl->pairs.erase({value_1, value_2});
 }
 
-void PairIntHashSet::for_each(std::function<void(int, int)> func) const {
+void PairIntHashSet::for_each(std::function<void(size_t, size_t)> func) const {
     for (const auto& [value_1, value_2] : m_impl->pairs)
         func(value_1, value_2);
 }
@@ -58,7 +58,7 @@ void PairIntHashSet::clear() { m_impl->pairs.clear(); }
 
 class I_IntPair_ToInt_HashMap {
   public:
-    std::unordered_map<std::pair<int, int>, int, int_pair_hash> map;
+    std::unordered_map<std::pair<size_t, size_t>, size_t, int_pair_hash> map;
 };
 
 IntPair_ToInt_HashMap::IntPair_ToInt_HashMap() {
@@ -71,16 +71,16 @@ IntPair_ToInt_HashMap::IntPair_ToInt_HashMap(IntPair_ToInt_HashMap&&) noexcept =
 
 IntPair_ToInt_HashMap& IntPair_ToInt_HashMap::operator=(IntPair_ToInt_HashMap&&) noexcept = default;
 
-void IntPair_ToInt_HashMap::add(int key_1, int key_2, int value) {
+void IntPair_ToInt_HashMap::add(size_t key_1, size_t key_2, size_t value) {
     assert(!has(key_1, key_2) && "IntPair_ToInt_HashMap::add: pair already exists");
     m_impl->map[{key_1, key_2}] = value;
 }
 
-int IntPair_ToInt_HashMap::get(int key_1, int key_2) const {
+size_t IntPair_ToInt_HashMap::get(size_t key_1, size_t key_2) const {
     return m_impl->map.at({key_1, key_2});
 }
 
-bool IntPair_ToInt_HashMap::has(int key_1, int key_2) const {
+bool IntPair_ToInt_HashMap::has(size_t key_1, size_t key_2) const {
     return m_impl->map.contains({key_1, key_2});
 }
 
@@ -88,7 +88,7 @@ size_t IntPair_ToInt_HashMap::size() const { return m_impl->map.size(); }
 
 bool IntPair_ToInt_HashMap::empty() const { return m_impl->map.empty(); }
 
-void IntPair_ToInt_HashMap::erase(int key_1, int key_2) {
+void IntPair_ToInt_HashMap::erase(size_t key_1, size_t key_2) {
     assert(has(key_1, key_2) && "IntPair_ToInt_HashMap::erase: pair does not exist");
     m_impl->map.erase({key_1, key_2});
 }
@@ -97,7 +97,7 @@ void IntPair_ToInt_HashMap::erase(int key_1, int key_2) {
 
 class I_IntHashSet {
   public:
-    std::unordered_set<int> ints;
+    std::unordered_set<size_t> ints;
 };
 
 IntHashSet::IntHashSet() { m_impl = std::make_unique<I_IntHashSet>(); }
@@ -108,11 +108,11 @@ IntHashSet::IntHashSet(IntHashSet&&) noexcept = default;
 
 IntHashSet& IntHashSet::operator=(IntHashSet&&) noexcept = default;
 
-void IntHashSet::add(int value) { m_impl->ints.insert(value); }
+void IntHashSet::add(size_t value) { m_impl->ints.insert(value); }
 
-bool IntHashSet::has(int value) const { return m_impl->ints.contains(value); }
+bool IntHashSet::has(size_t value) const { return m_impl->ints.contains(value); }
 
-int IntHashSet::get_one_int() const {
+size_t IntHashSet::get_one_int() const {
     assert(!empty() && "IntContainer::get_one_int: container is empty");
     return *m_impl->ints.begin();
 }
@@ -121,13 +121,13 @@ size_t IntHashSet::size() const { return m_impl->ints.size(); }
 
 bool IntHashSet::empty() const { return m_impl->ints.empty(); }
 
-void IntHashSet::erase(int value) {
+void IntHashSet::erase(size_t value) {
     assert(has(value) && "IntContainer::erase: value does not exist");
     m_impl->ints.erase(value);
 }
 
-void IntHashSet::for_each(std::function<void(int)> func) const {
-    for (int id : m_impl->ints)
+void IntHashSet::for_each(std::function<void(size_t)> func) const {
+    for (size_t id : m_impl->ints)
         func(id);
 }
 
@@ -135,7 +135,7 @@ void IntHashSet::for_each(std::function<void(int)> func) const {
 
 class I_Int_ToIntContainer_HashMap {
   public:
-    std::unordered_map<int, IntHashSet> key_values;
+    std::unordered_map<size_t, IntHashSet> key_values;
 };
 
 Int_ToIntContainer_HashMap::Int_ToIntContainer_HashMap() {
@@ -150,31 +150,31 @@ Int_ToIntContainer_HashMap::Int_ToIntContainer_HashMap(Int_ToIntContainer_HashMa
 Int_ToIntContainer_HashMap&
 Int_ToIntContainer_HashMap::operator=(Int_ToIntContainer_HashMap&&) noexcept = default;
 
-void Int_ToIntContainer_HashMap::add(int key, int value) {
+void Int_ToIntContainer_HashMap::add(size_t key, size_t value) {
     assert(!has(key, value) && "Int_ToIntContainer_HashMap::add: key-value already exists");
     m_impl->key_values[key].add(value);
 }
 
-bool Int_ToIntContainer_HashMap::has(int key, int value) const {
+bool Int_ToIntContainer_HashMap::has(size_t key, size_t value) const {
     if (!m_impl->key_values.contains(key))
         return false;
     return m_impl->key_values.at(key).has(value);
 }
 
-const IntHashSet& Int_ToIntContainer_HashMap::get(int node_id) const {
+const IntHashSet& Int_ToIntContainer_HashMap::get(size_t node_id) const {
     return m_impl->key_values[node_id];
 }
 
-void Int_ToIntContainer_HashMap::erase(int key, int value) {
+void Int_ToIntContainer_HashMap::erase(size_t key, size_t value) {
     assert(has(key, value) && "Int_ToIntContainer_HashMap::erase: key-value does not exist");
     m_impl->key_values[key].erase(value);
 }
 
-void Int_ToIntContainer_HashMap::erase(int node_id) { m_impl->key_values.erase(node_id); }
+void Int_ToIntContainer_HashMap::erase(size_t node_id) { m_impl->key_values.erase(node_id); }
 
 class I_Int_ToInt_HashMap {
   public:
-    std::unordered_map<int, int> map;
+    std::unordered_map<size_t, size_t> map;
 };
 
 Int_ToInt_HashMap::Int_ToInt_HashMap() { m_impl = std::make_unique<I_Int_ToInt_HashMap>(); }
@@ -185,24 +185,24 @@ Int_ToInt_HashMap::Int_ToInt_HashMap(Int_ToInt_HashMap&&) noexcept = default;
 
 Int_ToInt_HashMap& Int_ToInt_HashMap::operator=(Int_ToInt_HashMap&&) noexcept = default;
 
-void Int_ToInt_HashMap::add(int key, int value) { m_impl->map[key] = value; }
+void Int_ToInt_HashMap::add(size_t key, size_t value) { m_impl->map[key] = value; }
 
-bool Int_ToInt_HashMap::has(int key) const { return m_impl->map.contains(key); }
+bool Int_ToInt_HashMap::has(size_t key) const { return m_impl->map.contains(key); }
 
-int Int_ToInt_HashMap::get(int key) const { return m_impl->map.at(key); }
+size_t Int_ToInt_HashMap::get(size_t key) const { return m_impl->map.at(key); }
 
-void Int_ToInt_HashMap::erase(int key) { m_impl->map.erase(key); }
+void Int_ToInt_HashMap::erase(size_t key) { m_impl->map.erase(key); }
 
-void Int_ToInt_HashMap::for_each(std::function<void(int, int)> func) const {
+void Int_ToInt_HashMap::for_each(std::function<void(size_t, size_t)> func) const {
     for (const auto& [value_1, value_2] : m_impl->map)
         func(value_1, value_2);
 }
 
 void Int_ToInt_HashMap::clear() { m_impl->map.clear(); }
 
-void Int_ToInt_HashMap::update(int key, int value) {
+void Int_ToInt_HashMap::update(size_t key, size_t value) {
     assert(has(key));
     m_impl->map.at(key) = value;
 }
 
-int& Int_ToInt_HashMap::operator[](int key) { return m_impl->map[key]; }
+size_t& Int_ToInt_HashMap::operator[](size_t key) { return m_impl->map[key]; }
